requirements.txt
discord.py==2.3.2
matplotlib==3.8.3
numpy==2.2.0
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
Pillow==10.1.0
Pillow==11.0.0
zhmiscellany==2.6.3
